% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.9.r
\name{compareGC}
\alias{compareGC}
\title{Compare growth curves of tumor volume between Arms}
\usage{
compareGC(data, compare.to = c('all','neg.control','pos.control'), 
          control, n = 1000, fun, adjust)
}
\arguments{
\item{data}{a data frame of measured tumor volume data.}

\item{compare.to}{the type of compare. "all" means comparing all pairs of arms, "neg.control" means comparing to negative control arm, "pos.control" means comparing to posative control arm.}

\item{control}{the control arm. While compare to "neg.control", the negative control arm; While compare to "pos.control", the posative control arm.}

\item{n}{the times of permuation.}

\item{fun}{funciton used to caculate the statistics for permuation,MeanT or MeanW.}

\item{adjust}{the method for adjust p.value.}
}
\value{
compareGC returns a list with two components, stat and p.value, containing the observed statistics and the estimated p-value. compareGrowthCurves returns a data frame with components
Arm1:	  name of first arm in a comparison;
Arm2:	  name of second arm in a comparison;
Stat:  	observed value of the statistic;
p.value:	estimated p-value;
q.value:	p-value adjusted for multiple testing.
}
\description{
Compare the growth curves of tumor volume between pairwise arms with permuation test.
}
\details{
compareGC does pairwise comparisons of growth curves between all pairwise arms, or compared to negative/posative arm . p-values can be obtained with permutation test by setting n to some large value, n = 1000.
}
\examples{
\donttest{
data(oneAN.volume.data)
## compare each pair of arms
compareGC(data = oneAN.volume.data, compare.to = 'all', n = 1000, 
          fun = MeanT, adjust = "BH")
compareGC(data = oneAN.volume.data, compare.to = 'all', n = 1000, 
          fun = MeanW, adjust = "BH")

## compare to negative control
compareGC(data = oneAN.volume.data, 
          compare.to = 'neg.control', control = 'Control', 
          n = 1000, fun = MeanT, adjust = "BH")

## compare to positive control
compareGC(data = oneAN.volume.data, 
          compare.to = 'pos.control', control = 'Treatment_1', 
          n = 1000, fun = MeanT, adjust = "BH")
}
  
}
\references{
Elso, C.M., et al. Leishmaniasis host response loci (lmr1-3) modify disease severity through a Th1/Th2-independent pathway. Genes and immunity 2004;5(2):93-100.
}
\seealso{
compareTwoGC, MeanT, MeanW
}
