% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.10.r
\name{compareTwoGC}
\alias{compareTwoGC}
\title{Compare growth curves of tumor volume between pairwise Arms}
\usage{
compareTwoGC(data, n = n, fun)
}
\arguments{
\item{data}{a data frame of measured tumor volume data.}

\item{n}{the times of permuation.}

\item{fun}{funciton used to caculate the statistics for permuation.}
}
\description{
compare the growth curve of tumor volume between pairwise arms with permuation test.
}
\references{
Elso, C.M., et al. Leishmaniasis host response loci (lmr1-3) modify disease severity through a Th1/Th2-independent pathway. Genes and immunity 2004;5(2):93-100.
}
\seealso{
\code{\link{compareGC}}
}
