% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.9.r
\name{DataSummary}
\alias{DataSummary}
\title{Data summary}
\usage{
DataSummary(data, type, pattern = c("oneAN", "TAN"), measure.var, group.vars)
}
\arguments{
\item{data}{a data frame of measured tumor vloume data or body weight data of mouse.}

\item{type}{the type of data, "Vomule" or "BodyWeight".}

\item{pattern}{the pattern of PDX trial design, "oneAN" or "TAN".}

\item{measure.var}{the measured variable, such as "Volume", "BodyWeight".}

\item{group.vars}{the group variables.}
}
\value{
A data frame including the sample size, mean, SD, SE of group variables.
}
\description{
Summarize the mean, SD, SE, and sample size of each arm in input data.
}
\examples{
## summary the tumor volume data of oneAN pattern
data(oneAN.volume.data)
oneAN.volume.data[1:10,]
oneAN.v.s <- DataSummary(data = oneAN.volume.data,
                         type = 'Volume',
                         pattern = 'oneAN',
                         measure.var = 'Volume',
                         group.vars = c('Arms','Times'))
head(oneAN.v.s)

## summary the body weight data of oneAN pattern
data(oneAN.bw.data)
oneAN.bw.data[1:10,]
oneAN.bw.s <-DataSummary(data = oneAN.bw.data,
                         type = 'BodyWeight',
                         pattern = 'oneAN',
                         measure.var = 'BodyWeight',
                         group.vars = c('Arms','Times'))
head(oneAN.bw.s)

\donttest{
## summary the tumor volume data of TAN pattern
data(TAN.volume.data)
TAN.volume.data[1:10,]
TAN.v.s <- DataSummary(data = TAN.volume.data,
                       type = 'Volume',
                       pattern = 'TAN',
                       measure.var = 'Volume',
                       group.vars = c('Tumor','Arms','Times'))
head(TAN.v.s)

## summary the body weight data of TAN pattern
data(TAN.bw.data)
TAN.bw.data[1:10,]
TAN.bw.s <- DataSummary(data = TAN.bw.data,
                        type = 'BodyWeight',
                        pattern = 'TAN',
                        measure.var = 'BodyWeight',
                        group.vars = c('Tumor','Arms','Times'))
head(TAN.bw.s)
}

}
