% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.11.r
\name{DRAnalysis}
\alias{DRAnalysis}
\title{Assess differences in tumor volume between arms}
\usage{
DRAnalysis (data, pattern = c('oneAN','TAN'),
                   method=c('endpoint.ANOVA','endpoint.KW','endpoint.SRH',
                            'GR.ANOVA','GR.KW','GR.SRH','mixed.ANOVA','LMM'))
}
\arguments{
\item{data}{a data frame of measured tumor volume data.}

\item{pattern}{the pattern of PDX trial design, "oneAN" or "TAN".}

\item{method}{the method used to analysis.}
}
\value{
The statistical analysis results of the choosen method.
}
\description{
Assess differences in tumor volume between arms.
}
\details{
DRAnalysis offers both parametic and non-parametic statistic methods to access whether the arms have significant effect on tumor volume. users nedd to choose the proper method based on the feature of data and the suitable conditions.
}
\examples{
###oneAN pattern
data(oneAN.volume.data)
DRAnalysis(oneAN.volume.data,pattern = 'oneAN',method = 'endpoint.ANOVA')
DRAnalysis(oneAN.volume.data,pattern = 'oneAN',method = 'endpoint.KW')
DRAnalysis(oneAN.volume.data,pattern = 'oneAN',method = 'GR.ANOVA')
DRAnalysis(oneAN.volume.data,pattern = 'oneAN',method = 'mixed.ANOVA')
DRAnalysis(oneAN.volume.data,pattern = 'oneAN',method = 'LMM')

###TAN pattern
data(TAN.volume.data)
DRAnalysis(TAN.volume.data,pattern = 'TAN',method = 'endpoint.ANOVA')
DRAnalysis(TAN.volume.data,pattern = 'TAN',method = 'endpoint.SRH')
DRAnalysis(TAN.volume.data,pattern = 'TAN',method = 'LMM')

}
