% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.9.r
\name{DRLevel}
\alias{DRLevel}
\title{Define the drug response level for each animal.}
\usage{
DRLevel(data, method = c('NPDXE.Response','PPTP.Response','RC.Response'), 
               criteria, neg.control, rm.neg.control=TRUE)
}
\arguments{
\item{data}{a data frame of measured volume data.}

\item{method}{the method used to quantify the drug response level. Currently available methods include NPDXE.Response (default),PPTP.Response,RC.Response.}

\item{criteria}{the criteria conrresponding to method.}

\item{neg.control}{the negative control arm.}

\item{rm.neg.control}{whether remove the negative control arm.}
}
\value{
The response level of each animal.
}
\description{
Define the drug response level of each animal based on tumor volume change.
}
\details{
Defining drug response level is the general pipeline in clinical trial and precinical animal trial. DRLevel offers three published ways to define response level in function DRLevel. Notably, the criteria of each way for defining response level could be adjusted by users based on the actual experimental data.
}
\examples{
### build the criteria

# NPDXE.criteria
npdxe.criteria <- data.frame(BestResponse.lower = c(-1000,-0.95,-0.5,0.35),
                             BestResponse.upper = c(-0.95,-0.5,0.35,1000),
                             BestAvgResponse.lower = c(-1000,-0.4,-0.2,0.3), 
                             BestAvgResponse.upper = c(-0.4,-0.2,0.3,1000), 
                             Level = c( 'CR','PR', 'SD','PD'))
npdxe.criteria
 
# PPTP.criteria 
pptp.criteria <- data.frame(min.RC.lower = c(-1,-1,-0.5,-0.5),
                            min.RC.upper = c(-0.5,-0.5,1000,1000),
                            min.Vol.lower = c(0,100,NA,NA),
                            min.Vol.upper = c(100,10000,NA,NA),
                            end.RC.lower =c(NA,NA,-1,0.25),
                            end.RC.upper = c(NA,NA,0.25,1000),
                            Level = c( 'CR','PR', 'SD','PD'))
pptp.criteria

# RC.criteria
rc.criteria <- data.frame(Response.lower = c(-1000,-0.35,0.35),
                          Response.upper = c(-0.35,0.35,1000),
                          Level = c( 'CR','SD','PD'))
rc.criteria

### drug response level
##
data(oneAN.volume.data)
oneAN.drl <- DRLevel(data = oneAN.volume.data, 
                     method = 'NPDXE.Response', 
                     criteria = npdxe.criteria, 
                     neg.control = 'Control')
oneAN.drl <- oneAN.drl[order(oneAN.drl$Arms),]
head(oneAN.drl)

\donttest{
##
data(TAN.volume.data)
TAN.drl <- DRLevel(data = TAN.volume.data, 
                   method = 'NPDXE.Response', 
                   criteria = npdxe.criteria, 
                   neg.control = 'Control')
head(TAN.drl)

##
data(TAone.volume.data)
head(TAone.volume.data)
TAone.drl <- DRLevel(data = TAone.volume.data, 
                     method = 'NPDXE.Response', 
                     criteria = npdxe.criteria, 
                     neg.control = 'Control')
head(TAone.drl)
}

}
\references{
Gao, H., et al. High-throughput screening using patient-derived tumor xenografts to predict clinical trial drug response. Nat Med 2015;21(11):1318-1325.
Murphy, B., et al. Evaluation of Alternative In Vivo Drug Screening Methodology: A Single Mouse Analysis. Cancer Res 2016;76(19):5798-5809.
Bertotti, A., et al. The genomic landscape of response to EGFR blockade in colorectal cancer. Nature 2015;526(7572):263-267.
}
\seealso{
NPDXEResponseLevel, PPTPResponseLevel, RCResponseLevel
}
