% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.11.r
\name{DRLevelSummary}
\alias{DRLevelSummary}
\title{Summary the drug response level}
\usage{
DRLevelSummary(data, by)
}
\arguments{
\item{data}{drug response level, the output of DRLevel.}

\item{by}{the group summaried by.}
}
\description{
Summary the frequency and propotion of animals at every drug response level in each group.
}
\examples{
# NPDXE.criteria
npdxe.criteria <- data.frame(BestResponse.lower = c(-1000,-0.95,-0.5,0.35),
                             BestResponse.upper = c(-0.95,-0.5,0.35,1000),
                             BestAvgResponse.lower = c(-1000,-0.4,-0.2,0.3), 
                             BestAvgResponse.upper = c(-0.4,-0.2,0.3,1000), 
                             Level = c( 'CR','PR', 'SD','PD'))
npdxe.criteria
                            
### oneAN pattern
data(oneAN.volume.data)
oneAN.drl <- DRLevel(data = oneAN.volume.data, 
                     method = 'NPDXE.Response', 
                     criteria = npdxe.criteria, 
                     neg.control = 'Control')
DRLevelSummary(oneAN.drl,by = 'Arms')

\donttest{
### TAN pattern
data(TAN.volume.data)
TAN.drl <- DRLevel(data = TAN.volume.data, 
                   method = 'NPDXE.Response', 
                   criteria = npdxe.criteria, 
                   neg.control = 'Control')
DRLevelSummary(TAN.drl,by = 'Arms')
DRLevelSummary(TAN.drl,by = c('Arms','Tumor'))

### TAone pattern
data(TAone.volume.data)
TAone.drl <- DRLevel(data = TAone.volume.data, 
                     method = 'NPDXE.Response', 
                     criteria = npdxe.criteria, 
                     neg.control = 'Control')
head(TAone.drl)
DRLevelSummary(data = TAone.drl,by=c('Arms','Type'))
}

}
