% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.9.r
\name{MeanT}
\alias{MeanT}
\title{Caculating mean t-statistic between two arms of tumor volume growth curves}
\usage{
MeanT(data.1, data.2)
}
\arguments{
\item{data.1}{a data frame of tumor volume for the first arm.}

\item{data.2}{a data frame of tumor volume for the second arm.}
}
\value{
mean t-statistic.
}
\description{
Caculating mean t-statistic between two arms of tumor volume growth curves for all time points.
}
\details{
This function computes the t-statistic of tumor volumes between two arms at each time point, and returns the mean of t-statistics at all time points. This function is used by compareGC.
}
\references{
Elso, C.M., et al. Leishmaniasis host response loci (lmr1-3) modify disease severity through a Th1/Th2-independent pathway. Genes and immunity 2004;5(2):93-100.
}
\seealso{
MeaW
}
