% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.10.r
\name{MeanW}
\alias{MeanW}
\title{Caculating mean Wilcox-statistic between two arms of tumor volume growth curves}
\usage{
MeanW(data.1, data.2)
}
\arguments{
\item{data.1}{a data frame of tumor volume for the first arm.}

\item{data.2}{a data frame of tumor volume for the second arm.}
}
\value{
mean W-statistic.
}
\description{
Caculating mean Wilcox-statistic between two arms of tumor volume growth curves for all time points.
}
\details{
This function computes the W-statistic of tumor volumes between two arms at each time point, and returns the mean of W-statistics at all time points. This function is used by compareGC.
}
\seealso{
MeaT
}
