% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.9.r
\name{NPDXEResponseLevel}
\alias{NPDXEResponseLevel}
\title{Define drug response level based on Novartis PDXE response criteria}
\usage{
NPDXEResponseLevel(data, criteria)
}
\arguments{
\item{data}{data frame of measured volume data.}

\item{criteria}{the criteria used to defined response level.}
}
\value{
The response level of each animal.
}
\description{
Define the drug response level of each animal based on Novartis PDXE response criteria.
}
\details{
The 'Reference' for suitable condition of NPDXE response cirteria is: The initional tumor volume is about 200 mm3.
}
\references{
Gao, H., et al. High-throughput screening using patient-derived tumor xenografts to predict clinical trial drug response. Nat Med 2015;21(11):1318-1325.
}
\seealso{
DRLevel, PPTPResponseLevel, RCResponseLevel
}
