\name{oneAN.bw.data}
\docType{data}
\alias{oneAN.bw.data}
\title{Example of body weight data for oneAN pattern}
\description{
An example of body weight data for oneAN pattern, which includes the required imformation for DRAP package.
}
\usage{
data(oneAN.bw.data)
}
\details{
Body weigth data for oneAN pattern must at least include the columns "Arms" "ID" "Times" and "BodyWeight".
}
\examples{
data(oneAN.bw.data)
oneAN.bw.data[1:10,]
}
\keyword{datasets}