% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.10.r
\name{plotBWeightGC}
\alias{plotBWeightGC}
\title{Plot growth curves for body weight of mouses}
\usage{
plotBWeightGC(data, level = c('Animal','Arm'), 
              pattern = c("OneAN", "TAN"), 
              orders = NULL, position.dodge)
}
\arguments{
\item{data}{a data frame of measured body weight data of mouse.}

\item{level}{the level to present, "Animal" level or "Arm" level.}

\item{pattern}{the pattern of PDX trial design, "OneAN" or "TAN".}

\item{orders}{the redefined order for visulization. For oneAN pattern,the order is redifined orders of the 'Arms'; for TAN pattern, the order is redifined orders of the 'Tumor'.}

\item{position.dodge}{Dodging preserves the vertical position of an geom while adjusting the horizontal position.}
}
\description{
Plot growth curves for body weight of mouses.
}
\examples{
### oneAN pattern
data(oneAN.bw.data)
plotBWeightGC(data = oneAN.bw.data,level = 'Animal',
              pattern = 'oneAN')
plotBWeightGC(data = oneAN.bw.data,level = 'Arm',
              pattern = 'oneAN',position.dodge = 0.5)

\donttest{
### TAN pattern
data(TAN.bw.data) 
plotBWeightGC(data = TAN.bw.data,level = 'Animal',
              pattern = 'TAN')
plotBWeightGC(data = TAN.bw.data,level = 'Arm',
              pattern = 'TAN',position.dodge = 0.5)
}

}
