% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.9.r
\name{plotDRLevel}
\alias{plotDRLevel}
\title{Visulize the drug response level}
\usage{
plotDRLevel(data, by = c("Arms", "Tumor"), 
           pattern = c("oneAN",'TAN'), orders=NULL, 
           orders.1=NULL, orders.2=NULL)
}
\arguments{
\item{data}{drug response level ,the output of DRLevel.}

\item{by}{the level summaried by.}

\item{pattern}{the pattern of PDX trial design, "oneAN" or "TAN".}

\item{orders}{the redefined order while the argument "by" just contain one variable.}

\item{orders.1}{the redefined order of the first varible while the argument "by" contain two variables.}

\item{orders.2}{the redefined order of the second varible while the argument "by" contain two variables.}
}
\description{
Visulize the drug response level for each arm.
}
\examples{
# NPDXE.criteria
npdxe.criteria <- data.frame(BestResponse.lower = c(-1000,-0.95,-0.5,0.35),
                             BestResponse.upper = c(-0.95,-0.5,0.35,1000),
                             BestAvgResponse.lower = c(-1000,-0.4,-0.2,0.3), 
                             BestAvgResponse.upper = c(-0.4,-0.2,0.3,1000), 
                             Level = c( 'CR','PR', 'SD','PD'))
npdxe.criteria
                            
### oneAN pattern
data(oneAN.volume.data)
oneAN.drl <- DRLevel(data = oneAN.volume.data, 
                     method = 'NPDXE.Response', 
                     criteria = npdxe.criteria, 
                     neg.control = 'Control')
DRLevelSummary(oneAN.drl,by = 'Arms')
plotDRLevel(data = oneAN.drl,by = 'Arms',pattern = 'oneAN')

\donttest{
### TAN pattern
data(TAN.volume.data)
TAN.drl <- DRLevel(data = TAN.volume.data, 
                   method = 'NPDXE.Response', 
                   criteria = npdxe.criteria, 
                   neg.control = 'Control')
DRLevelSummary(TAN.drl,by = 'Arms')
plotDRLevel(data = TAN.drl, by='Arms', pattern = 'TAN')
plotDRLevel(data = TAN.drl, by=c('Arms','Tumor'), pattern = 'TAN')

# rectify orders
arms <- unique(TAN.volume.data$Arms)
arms <- arms[-1]
arms
tumors <- unique(TAN.volume.data$Tumor)
tumors
plotDRLevel(data = TAN.drl, by = c('Arms','Tumor'), 
            pattern = 'TAN', orders.1 = arms, orders.2 = tumors)
}

}
