% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.10.r
\name{plotDRResults}
\alias{plotDRResults}
\title{plot the results of drug response level analysis.}
\usage{
plotDRResults(data, by, min.number, measurement = c('RR','DCR','both'))
}
\arguments{
\item{data}{drug response level ,the output of DRLevel.}

\item{by}{the level summaried by.}

\item{min.number}{the lower cutoff of sample number of every arm}

\item{measurement}{the response evaluation indexes used,  whether "RR", "DCR", or "both".}
}
\description{
Visulize the result of drug response analysis, such as RR or DCR of each arm.
}
\examples{
# NPDXE.criteria 
npdxe.criteria <- data.frame(BestResponse.lower = c(-1000,-0.95,-0.5,0.35),
                             BestResponse.upper = c(-0.95,-0.5,0.35,1000),
                             BestAvgResponse.lower = c(-1000,-0.4,-0.2,0.3), 
                             BestAvgResponse.upper = c(-0.4,-0.2,0.3,1000), 
                             Level = c( 'CR','PR', 'SD','PD'))
npdxe.criteria

data(TAN.volume.data)
TAN.drl <- DRLevel(data = TAN.volume.data, 
                   method = 'NPDXE.Response', 
                   criteria = npdxe.criteria, 
                   neg.control = 'Control')
plotDRResults(TAN.drl,by = 'Arms',measurement = 'DCR',min.number = 5)

}
