% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.9.r
\name{plotEndpoint}
\alias{plotEndpoint}
\title{visulize the tumor volume data measured at the end of experiment}
\usage{
plotEndpoint(data, pattern, orders = NULL,type = c('barplot','boxplot'))
}
\arguments{
\item{data}{a data frame of measured tumor volume data.}

\item{pattern}{the pattern of PDX trial design, "oneAN" or "TAN".}

\item{orders}{the redefined order for visulization. For oneAN pattern,the order is redifined orders of the 'Arms'; for TAN pattern, the order is redifined orders of the 'Tumor'.}

\item{type}{the type to visulize, "barplot" or "boxplot".}
}
\description{
visulize the tumor volume data measured at the end of experiment.
}
\examples{
### oneAN pattern
data(oneAN.volume.data)
plotEndpoint(oneAN.volume.data,type = 'boxplot',pattern = 'oneAN')
plotEndpoint(oneAN.volume.data,type = 'barplot',pattern = 'oneAN')

\donttest{
### TAN pattern
data(TAN.volume.data)
plotEndpoint(TAN.volume.data,pattern = 'TAN',type = 'barplot')
plotEndpoint(TAN.volume.data,pattern = 'TAN',type = 'boxplot')
}

}
