% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.9.r
\name{plotGR}
\alias{plotGR}
\title{visulize the growth rate of tumor volume}
\usage{
plotGR(data, pattern, orders = NULL, type = c('barplot','boxplot'))
}
\arguments{
\item{data}{a data frame of measured tumor volume data.}

\item{pattern}{the pattern of PDX trial design, "oneAN" or "TAN".}

\item{orders}{the redefined order for visulization. For oneAN pattern,the order is redifined orders of the 'Arms';for TAN pattern, the order is redifined orders of the 'Tumor'.}

\item{type}{the type to visulize, "barplot" or "boxplot".}
}
\description{
visulize the growth rate of tumor volume.
}
\examples{
### oneAN pattern
data(oneAN.volume.data)
plotGR(oneAN.volume.data,type = 'boxplot',pattern = 'oneAN')
plotGR(oneAN.volume.data,type = 'barplot',pattern = 'oneAN')

\donttest{
### TAN pattern
data(TAN.volume.data)
plotGR(TAN.volume.data,pattern = 'TAN',type = 'barplot')
plotGR(TAN.volume.data,pattern = 'TAN',type = 'boxplot')
}

}
