% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.11.r
\name{plotRC}
\alias{plotRC}
\title{plot the relative change of data}
\usage{
plotRC(data, type, pattern, rm.baseline = TRUE, orders = NULL, ...)
}
\arguments{
\item{data}{a data frame of measured tumor volume data or body weight data.}

\item{type}{the type of data, "Vomule"  or "BodyWeight".}

\item{pattern}{the pattern of PDX trial design, "oneAN" or "TAN".}

\item{rm.baseline}{a logical parameter, whether remove the baseline.}

\item{orders}{the redefined order for visulization. For oneAN pattern,the order is redifined orders of the 'Arms'; for TAN pattern, the order is redifined orders of the 'Tumor'.}

\item{...}{other parameters passed to ggplot.}
}
\description{
plot the relative change of tumor volume data or body weight data.
}
\examples{
### oneAN pattern
data(oneAN.volume.data)
plotRC(data = oneAN.volume.data,type = 'Volume',pattern = 'oneAN')

data(oneAN.bw.data)
plotRC(data = oneAN.bw.data,type = 'BodyWeight',pattern = 'oneAN')

\donttest{
### TAN pattern
data(TAN.volume.data)
plotRC(data = TAN.volume.data,type = 'Volume',pattern = 'TAN')

data(TAN.bw.data)
plotRC(data = TAN.bw.data,type = 'BodyWeight',pattern = 'TAN')
}

}
