% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.9.r
\name{plotTGI}
\alias{plotTGI}
\title{TGI visulizing}
\usage{
plotTGI(data, pattern, scope = c('end.point','all.point'),orders=NULL)
}
\arguments{
\item{data}{the output of function TGI.}

\item{pattern}{the pattern of PDX trial design, "oneAN" or "TAN".}

\item{scope}{the scope of time point to visulize, whether "end,point" or "all.point".}

\item{orders}{the redefined order for visulization. For oneAN pattern,the order is redifined orders of the 'Arms'; for TAN pattern, the order is redifined orders of the 'Tumor'.}
}
\description{
Visulize TGI of each Arms.
}
\examples{
### oneAN pattern
data(oneAN.volume.data)
oneAN.tgi <- TGI(data = oneAN.volume.data,neg.control = 'Control',method = 'AUC',pattern = 'oneAN')
plotTGI(oneAN.tgi,pattern = 'oneAN',scope = 'end.point')
plotTGI(oneAN.tgi,pattern = 'oneAN',scope = 'all.point')

\donttest{
### TAN pattern
data(TAN.volume.data)
TAN.tgi <- TGI(TAN.volume.data,  neg.control = 'Control',pattern = 'TAN',method = 'AUC')
plotTGI(TAN.tgi,pattern = 'TAN',scope = 'end.point')
plotTGI(TAN.tgi,pattern = 'TAN',scope = 'all.point')
}

}
