% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.9.r
\name{plotVolumeGC}
\alias{plotVolumeGC}
\title{Plot growth curves of tumor volume data}
\usage{
plotVolumeGC(data, level = c('Animal','Arm'), 
             pattern = c("OneAN", "TAN"), orders = NULL, 
             position.dodge)
}
\arguments{
\item{data}{a data frame of measured tumor volume data.}

\item{level}{the level to present, "Animal" level or "Arm" level.}

\item{pattern}{the pattern of PDX trial design, "OneAN" or "TAN".}

\item{orders}{the redefined order for visulization. For oneAN pattern,the order is redifined orders of the 'Arms'; for TAN pattern, the order is redifined orders of the 'Tumor'.}

\item{position.dodge}{Dodging preserves the vertical position of an geom while adjusting the horizontal position.}
}
\description{
Plot growth curves of tumor volume data.
}
\examples{

### oneAN pattern
data(oneAN.volume.data)
# Animal level
plotVolumeGC(data = oneAN.volume.data,
             level = 'Animal',pattern = 'oneAN')

# Arm level
plotVolumeGC(data = oneAN.volume.data,level = 'Arm',
             pattern = 'oneAN',position.dodge = 0.5)

## reset order of Arms
orders <- unique(oneAN.volume.data$Arms)
orders <- orders[c(2:6,1)]
plotVolumeGC(data = oneAN.volume.data,level = 'Animal',
             pattern = 'oneAN',orders = orders)

\donttest{
### TAN pattern
data(TAN.volume.data)
# Animal level
plotVolumeGC(data = TAN.volume.data,level = 'Animal',
             pattern = 'TAN')

# Arm level
plotVolumeGC(data = TAN.volume.data,level = 'Arm',
             pattern = 'TAN',position.dodge = 0.5)

## reset order of tumors
orders <- unique(TAN.volume.data$Tumor)
orders <- orders[c(2:4,1)]
plotVolumeGC(data = TAN.volume.data,level = 'Animal',
             pattern = 'TAN',orders = orders)
plotVolumeGC(data = TAN.volume.data,level = 'Arm',
             pattern = 'TAN',orders = orders,position.dodge = 0.5)
}

}
