% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.10.r
\name{plotWaterfall}
\alias{plotWaterfall}
\title{plot waterfall}
\usage{
plotWaterfall(data, max.threshold = 100, order.by = c("Best.Avg.Response",
  "Best.Response", "RC.Response"))
}
\arguments{
\item{data}{drug response level, the output of DRLevel.}

\item{max.threshold, }{the max threshold of volume change meseaure to visulize.}

\item{order.by}{the meseasure ordered by.}
}
\description{
Visulize the drug response for each animal or each Tumor in the form of waterfall.
}
\examples{
\donttest{
## NPDXE.criteria
npdxe.criteria <- data.frame(BestResponse.lower = c(-1000,-0.95,-0.5,0.35),
                             BestResponse.upper = c(-0.95,-0.5,0.35,1000),
                             BestAvgResponse.lower = c(-1000,-0.4,-0.2,0.3), 
                             BestAvgResponse.upper = c(-0.4,-0.2,0.3,1000), 
                             Level = c( 'CR','PR', 'SD','PD'))
npdxe.criteria

data(TAone.volume.data)
TAone.drl <- DRLevel(data = TAone.volume.data, 
                     method = 'NPDXE.Response', 
                     criteria = npdxe.criteria, 
                     neg.control = 'Control')

# Choose the data of GC responsed to BYL719 + LJM716
bl.gc.drl <- TAone.drl[TAone.drl$Arms == "BYL719 + LJM716" & TAone.drl$Type == 'GC',]
plotWaterfall(data = bl.gc.drl, max.threshold = 100,order.by = 'Best.Response')
}

}
