% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.9.r
\name{RCResponseLevel}
\alias{RCResponseLevel}
\title{Define response level based on relatvie change of tumor volume}
\usage{
RCResponseLevel(data, criteria)
}
\arguments{
\item{data}{a data frame of measured volume data.}

\item{criteria}{the criteria used to defined response level.}
}
\value{
The response level of each animal.
}
\description{
Define the drug response level of each animal based on the relative change of tumor volume.
}
\details{
The 'Reference' for suitable condition of RC response criteria includes: 1. The initional tumor volume is about 400 mm3; 2. Drug treatment lasts to 21 days.
}
\references{
Bertotti, A., et al. The genomic landscape of response to EGFR blockade in colorectal cancer. Nature 2015;526(7572):263-267.
}
\seealso{
DRLevel, NPDXEResponseLevel, PPTPResponseLevel
}
