% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.10.r
\name{RelativeChange}
\alias{RelativeChange}
\title{Caculate the relative change for tumor volume data or body weight data.}
\usage{
RelativeChange(data,type,rm.baseline = TRUE)
}
\arguments{
\item{data}{a data frame of measured tumor vloume data or body weight data of mouse.}

\item{type}{the type of data, "Vomule" or "BodyWeight".}

\item{rm.baseline}{a logical parameter, whether remove the baseline.}
}
\value{
If rm.baseline = T, it returns the relative change of tumor volume data or body weight data; if rm.baseline = F, it returns the fold change of tumor volume data or body weight data.
}
\description{
Caculate the relative change for tumor volume data or body weight data.
}
\details{
If rm.baseline = T, the way of caculating follows like Vt/V0 - 1; if rm.baseline = F, the way of caculating follows like Vt/V0.
}
\examples{
data(oneAN.volume.data)
## caculate relative change of tumor volume
oneAN.v.rc <- RelativeChange(oneAN.volume.data,type = 'Volume',rm.baseline = TRUE)
head(oneAN.v.rc)

##caculate fold change of tumor volume
oneAN.v.fc <- RelativeChange(oneAN.volume.data,type = 'Volume',rm.baseline = FALSE)
head(oneAN.v.fc)

##caculate relative change of body weight data
data(oneAN.bw.data)
oneAN.bw.rc <- RelativeChange(oneAN.bw.data,type = 'BodyWeight',rm.baseline = TRUE)
head(oneAN.bw.rc)

}
