% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAP_code_1.5.9.r
\name{TGI}
\alias{TGI}
\title{Tumor growth inhibition (TGI) rate caculating}
\usage{
TGI(data, neg.control, pattern = c('oneAN','TAN'), method = c('deltaV','RCV','AUC'))
}
\arguments{
\item{data}{a data frame of measured tumor volume data.}

\item{neg.control}{the negative control arm.}

\item{pattern}{the pattern of PDX trial design, "oneAN" or "TAN".}

\item{method}{the method used to caculate TGI.}
}
\value{
The TGI value of each experimental arm.
}
\description{
Caculate tumor growth inhibition (TGI) rate of each Arms based on negative control arm.
}
\note{
Tumor growth inhibition (TGI) is one of the most widely used indicators to mesure the effect of tumor growth inhibition of treatment. Even though many forms of caculating TGI have been apllied in publications and the field of drug R&D, the basic form is following: 
  TGI = (1 - F(VT)/F(VC)). 
F(VT) and F(VC) mean the calculating ways for the treatment arm and control arm respectively. DRAP provides three types of F( ) function to calculate TGI: (1) F=Vt-V0; (2) F=Vt/V0; (3) F = area under the curve of tumor volume. Vt and V0 represent the mean tumor volume at the time t and time 0 respectively.
}
\examples{
### oenAN pattern
data(oneAN.volume.data)
oneAN.tgi <- TGI(data = oneAN.volume.data,
                 neg.control = 'Control',
                 method = 'AUC',
                 pattern = 'oneAN')

\donttest{
###TAN pattern
data(TAN.volume.data)
TAN.tgi <- TGI(TAN.volume.data,
               neg.control = 'Control',
               pattern = 'TAN',
               method = 'AUC')
}

}
